+++ Changelog.md
@@ -26,6 +26,18 @@ HOW I DO VERSIONING:
 <!-- ### :rocket: Performance Improvements -->
 
 
+## [3.1.3] - 2026/03/27
+<!-- ### :warning: Known issues -->
+<!-- ### :ballot_box_with_check: Fixed -->
+### :heavy_exclamation_mark: Changed
+- Optimized hmax/hadd greedy choice
+- Using watch preconditions in greedy algorithm
+<!-- ### :heavy_plus_sign: Added -->
+<!-- ### :x: Removed -->
+<!-- ### :curly_loop: Other -->
+<!-- ### :rocket: Performance Improvements -->
+
+
 ## [3.1.2] - 2026/03/26
 ### :warning: Known issues
 - Execution on MacOS is (sometimes) non-deterministic (this is a CPLEX bug for ARM MacOS)